Simplify test file server with httptest
